+++ CHANGELOG.md
@@ -56,7 +56,7 @@ Current
 
 ### Added:
 - [Add ability in fili-sql to translate FilteredAggregation into SQL](https://github.com/yahoo/fili/pull/1083)
-   * Translate a Druid query with `n` FilteredAggregation into SQL using `\(n + 1\)` subquery unions. 
+   * Translate a Druid query with `n` FilteredAggregation into SQL using `(n + 1)` subquery unions. 
        - See PR description for details  
 
 - [Add ability to convert `TimeSeriesQuery` to `GroupByQuery`](https://github.com/yahoo/fili/issues/1059)